Move KTime / GetKTime from util to times package
